Refatora acrescentando retorno para tipo correto para ResultViewModel

+++ Manutec.Application/Commands/CustomerEntity/DeleteCustomerHandler.cs
@@ -1,9 +1,10 @@
 ﻿using Manutec.Application.Models;
+using Manutec.Application.Models.CustomerModel;
 using Manutec.Core.Repositories;
 using MediatR;
 
 namespace Manutec.Application.Commands.CustomerEntity;
-public class DeleteCustomerHandler : IRequestHandler<DeleteCustomerCommand, ResultViewModel>
+public class DeleteCustomerHandler : IRequestHandler<DeleteCustomerCommand, ResultViewModel<CustomerViewModel>>
 {
     private readonly ICustomerRepository _customerRepository;
 
@@ -11,17 +12,24 @@ public class DeleteCustomerHandler : IRequestHandler<DeleteCustomerCommand, Resu
     {
         _customerRepository = customerRepository;
     }
-    public async Task<ResultViewModel> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
+    public async Task<ResultViewModel<CustomerViewModel>> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
     {
         var customer = await _customerRepository.GetById(request.WorkShopId, request.Id);
 
         if (customer is null)
         {
-            return ResultViewModel.Error("Cliente não encontrado");
+            return ResultViewModel<CustomerViewModel>.Error("Cliente não encontrado");
+        }
+
+        if (customer.IsDeleted)
+        {
+            return ResultViewModel<CustomerViewModel>.Error("Cliente já foi excluído.");
         }
 
         await _customerRepository.Delete(customer);
 
-        return ResultViewModel.Success();
+        var model = CustomerViewModel.FromEntity(customer);
+
+        return ResultViewModel<CustomerViewModel>.Success(model);
     }
 }
